[KOGITO-7081] Updating example to work with quarkiverse
