+++ kogito-quarkus-examples/serverless-workflow-qas-service-showcase/query-answer-service/src/main/resources/application.properties
@@ -1,4 +1,5 @@
 quarkus.log.category."org.acme".level=DEBUG
 # OpenApi client properties to access the query-service
-quarkus.rest-client."org.openapi.quarkus.api.DefaultApi".url=${QUERYSERVICE_URL:http://localhost:8283}
+quarkus.rest-client."org.kie.kogito.openapi.queryservice.api.DefaultApi".url=${QUERYSERVICE_URL:http://localhost:8283}
 quarkus.swagger-ui.always-include=true
+quarkus.devservices.enabled=false
