+++ kogito-quarkus-examples/serverless-workflow-temperature-conversion/conversion-workflow/pom.xml
@@ -83,6 +83,8 @@
         <executions>
           <execution>
             <goals>
+              <goal>generate-code</goal>
+              <goal>generate-code-tests</goal>
               <goal>build</goal>
             </goals>
           </execution>
