+++ kogito-quarkus-examples/serverless-workflow-temperature-conversion/conversion-workflow/src/main/resources/application.properties
@@ -4,5 +4,8 @@
 quarkus.swagger-ui.always-include=true
 mp.openapi.extensions.smallrye.operationIdStrategy=METHOD
 # OpenApi Client Properties
-org.kogito.openapi.client.multiplication.base_path=http://localhost:8282
-org.kogito.openapi.client.subtraction.base_path=http://localhost:8181
+#org.kogito.openapi.client.multiplication.base_path=http://localhost:8282
+#org.kogito.openapi.client.subtraction.base_path=http://localhost:8181
+
+quarkus.rest-client."org.kie.kogito.openapi.subtraction.api.DefaultApi".url=http://localhost:8181
+quarkus.rest-client."org.kie.kogito.openapi.multiplication.api.DefaultApi".url=http://localhost:8282
