+++ kogito-quarkus-examples/serverless-workflow-temperature-conversion/conversion-workflow/src/main/resources/specs/subtraction.yaml
@@ -15,14 +15,19 @@ paths:
       responses:
         "200":
           description: OK
+          content: 
+            application/json:
+              schema: 
+                type: object
+                properties:
+                 difference:
+                   format: float
+                   type: number 
 components:
   schemas:
     SubtractionOperation:
       type: object
       properties:
-        difference:
-          format: float
-          type: number
         leftElement:
           format: float
           type: number
