+++ kogito-quarkus-examples/serverless-workflow-temperature-conversion/conversion-workflow/src/test/java/org/kie/kogito/serverless/OperationsMockService.java
@@ -38,10 +38,10 @@ public class OperationsMockService implements QuarkusTestResourceLifecycleManage
     public Map<String, String> start() {
         multiplicationService =
                 this.startServer(8282,
-                        "{\"multiplication\": { \"leftElement\": \"68.0\", \"rightElement\": \"0.5556\", \"product\": \"37.808\" }}");
+                        "{  \"product\": 37.808 }");
         subtractionService =
                 this.startServer(8181,
-                        "{\"subtraction\": { \"leftElement\": \"100\", \"rightElement\": \"32\", \"difference\": \"68.0\" }}");
+                        "{ \"difference\": 68.0 }");
         return Collections.emptyMap();
     }
